fix: out of order forks, record exec args
This addresses a problem where the timestamp of a fork would be _later_
than the timestamp of an exec. The root cause was that the timestamp was
derived from the `elapsed` variable on the `exit` probe of the system
call. This fix stores the `elapsed` value at the beginning of the system
call.

This also creates a `sort` command that sorts the events in a recording
by the timestamp of the event. A user shouldn't need this unless they
want to look at the raw recording. It's mostly as a debugging tool for
devs (i.e. me).

Finally, this creates an event for recording the arguments from an exec.
Previously this wasn't working for some reason (`join(args.argv)` would
simply print an empty string). That is still the case for some execs,
but most of them now print their arguments as a single string. I don't
know why it wasn't working before (neither did a `bpftrace` maintainer)
and I don't know why it's working now. This does mean that you lose the
exact quoting of the original arguments and can't tell how exactly they
were parsed into arguments, but I think it's more useful to have the
args than not. Right now there's duplicated work that reads procfs to
get the args, so maybe the arguments could be an enum where one variant
gets the args from procfs and keeps argv, and if the arguments couldn't
be found from procfs then the joined string of the exec args will be
used.

There's still currently an issue I'm seeing where the initial exec of
`flox activate` isn't being recorded, and it seems like it's also
missing the watchdog.