fix: truncate output files
Previously a file specified for output would simply write to the file
starting at the beginning, so if we didn't write past the end of the
previous contents of the file we would leave some of the previous
contents of the file behind. Now we truncate the file so that all of the
previous contents are removed before writing.