fix(script): change timestamps to elapsed
Previously I changed the timestamps from `nsecs` to `elapsed` and was
seeing that the forks and execs were out of order. In an attempt to
bring those into the correct order I switched from `elapsed` to
`nsecs(monotonic)`. Now what I'm seeing is that some events near the end
of execution have timestamps that are two orders of magnitude smaller
than earlier events, which obviously causes bugs when you rely on later
events having larger timestamps. To address that I've switched back to
`elapsed` for the time being until I find a different solution.