feat: correctly report execs
Previously proctrace would report all `exec` calls, regardless of
whether they succeeded or not. This generated excess output that was
confusing to read. Any system calls that fail are now filtered out.
Additionally, we also now report the filename of the program that is
being `exec`ed, which provides additional context that was previously
missing sometimes.

On the implementation level, events now have a sequence number. This
was important for correctly ordering events related to `exec`s. The
method used to print the filename is apparently slow enough that the
exit probe can fire before it's done, so you'll see the EXEC line
before you see the EXEC_ARGS line. Buffers are now sorted according
to this sequence number instead of the timestamp since events can also
occur close enough together that they have the same timestamp, and the
binary search used to order buffers will always stop at the *first*
event with a given timestamp, which causes events to be ordered
incorrectly.